#20 Implemented Full Override
Contains a bit of a sloppy hack in making a copy
X2Ability_DefaultAbilitySet.FinalizeHackAbility_BuildGameState(), which
will do until I can de-privatize the function in the Community
Highlander. Also there is only one hack reward, control robot for the
entire mission; the resistance MEC reward does not make sense in the
context of the base game.